creating separated class for enqueuing assets
