+++ functions.php
@@ -10,6 +10,10 @@ if( ! defined( 'LCODE_DIR_PATH' ) ) {
     define( 'LCODE_DIR_PATH', untrailingslashit( get_template_directory() ) );
 }
 
+if( ! defined( 'LCODE_DIR_URI' ) ) {
+    define( 'LCODE_DIR_URI', untrailingslashit( get_template_directory_uri() ) );
+}
+
 require_once LCODE_DIR_PATH . '/inc/helpers/autoloader.php';
 
 function lcode_get_theme_instance() {
